refactor(mem_cache): review cleanup — overflow guards, docs, lock-ref helpers
- Drop unreachable u32 lock_ref overflow/underflow guards (plain +/-1;
  debug builds still catch over/underflow via Rust arithmetic checks).
- Trim Component-trait docs to first sentence + contract (drop dispatch-order
  impl details); generalize postprocess_* docs; refresh stale 'LRU slot' docs.
- Move inc/dec_lock_ref_non_full from the Slot trait to utils.rs as free fns
  (like evict_non_full); revert evict_utils.rs -> utils.rs.
- cargo fmt.
