+++ rust/sglang-mem-cache/src/components/full.rs
@@ -1,8 +1,8 @@
 //! Tree node component for device FULL values.
 
 use super::{Component, IncLockRefResult, MatchValidator};
-use crate::component_type::ComponentType;
 use super::{EvictRequest, EvictResult, Slot, evict_full};
+use crate::component_type::ComponentType;
 use crate::tree_node_pool::{ChildKeyType, NodeIdx, TreeNode, TreeNodePool};
 
 #[derive(Default)]
@@ -81,7 +81,7 @@ impl<K: ChildKeyType> Component<K> for FullComponent {
     }
 }
 
-/// LRU for FULL values.
+/// FULL component slot.
 pub struct FullSlot;
 
 impl Slot for FullSlot {
@@ -100,23 +100,12 @@ impl Slot for FullSlot {
 
     /// Debit the parent's `num_children_with_device_full`.
     fn postprocess_take_value<K: ChildKeyType>(pool: &mut TreeNodePool<K>, parent_idx: NodeIdx) {
-        let count = &mut pool.get_mut(parent_idx).num_children_with_device_full;
-        #[allow(clippy::expect_used, reason = "underflow = dec without matching inc")]
-        let new = count
-            .checked_sub(1)
-            .expect("num_children_with_device_full underflow");
-        *count = new;
+        pool.get_mut(parent_idx).num_children_with_device_full -= 1;
     }
 
     fn inc_lock_ref<K: ChildKeyType>(pool: &mut TreeNodePool<K>, node_idx: NodeIdx) -> i64 {
         let node = pool.get_mut(node_idx);
-        #[allow(
-            clippy::expect_used,
-            reason = "u32 lock_ref overflow effectively impossible"
-        )]
-        let new = Self::lock_ref(node)
-            .checked_add(1)
-            .expect("FullSlot::inc_lock_ref: u32 overflow");
+        let new = Self::lock_ref(node) + 1;
         Self::set_lock_ref(node, new);
         if new != 1 {
             return 0;
@@ -131,10 +120,7 @@ impl Slot for FullSlot {
     fn dec_lock_ref<K: ChildKeyType>(pool: &mut TreeNodePool<K>, node_idx: NodeIdx) -> i64 {
         let node = pool.get_mut(node_idx);
         let component_idx = Self::COMPONENT as usize;
-        #[allow(clippy::expect_used, reason = "underflow = dec without matching inc")]
-        let new = Self::lock_ref(node)
-            .checked_sub(1)
-            .expect("FullSlot::dec_lock_ref: underflow");
+        let new = Self::lock_ref(node) - 1;
         // FULL's lock_ref must stay >= every other component's.
         assert!(
             node.components
