+++ rust/sglang-mem-cache/src/radix_cache_wrapper.rs
@@ -7,9 +7,9 @@ use pyo3::types::PyList;
 use crate::py_interop::PyTensor;
 
 use crate::component_type::NUM_COMPONENT_TYPES;
+use crate::components::{EvictRequest, EvictResult};
 use crate::deferred_action::DeferredAction;
 use crate::radix_cache::{BigramRadixCache, InsertResult, MatchResult, PageRadixCache};
-use crate::components::{EvictRequest, EvictResult};
 use crate::utils::parse_device;
 
 /// Per-component eviction outcome at the PyO3 boundary.
